refactor: rename Entry.js -> parse-and-generate.js
