+++ packages/compile/src/index.js
@@ -3,6 +3,6 @@
 export { canonicalName } from './_shared/helpers.js'
 export { readDat } from './_shared/read-dat.js'
 export { generateCode } from './generate/code-gen.js'
-export { parseAndGenerate, printNames } from './Entry.js'
+export { parseAndGenerate, printNames } from './parse-and-generate.js'
 export { parseModel } from './Parser.js'
 export { preprocessModel } from './preprocess/preprocessor.js'
